Consolidate some OpenCL kernels into one ERB template

+++ lib/tensor_stream/evaluator/kernels/abs.cl
@@ -7,10 +7,10 @@ __kernel void abs_fp(const int M, const int N, __global const float *A, __global
     C[globalRow * N + globalCol] = fabs(A[globalRow * N + globalCol]);
 }
 
-__kernel void abs_int(const int M, const int N, __global const float *A, __global float *C) {
+__kernel void abs_int(const int M, const int N, __global const int *A, __global int *C) {
     // Get the index of the current element to be processed
     const int globalRow = get_global_id(0); // Row ID of C (0..M)
     const int globalCol = get_global_id(1); // Col ID of C (0..N)
 
-    C[globalRow * N + globalCol] = fabs(A[globalRow * N + globalCol]);
+    C[globalRow * N + globalCol] = fabs((float)A[globalRow * N + globalCol]);
 }
